requirements.txt
mistralai>=1.0.0
python-dotenv>=1.0.0
pyyaml>=6.0
genanki>=0.13.0
requests>=2.31.0
libsass>=0.23.0
